Fix the fire API
Apparently minecraft can't toggle fire, only set how many ticks a fire will be active, so we publicly document the fact that we're just setting the entity to be on fire for Integer.MAX_VALUE ticks.
Since plugins will probably want to make eternaly-burning entities, add an API to make entities immune to fire.
Also, apparently fire belongs in Entity and not LivingEntity.