Use buildscript instead of gradle plugin DSL
Gradle plugin DSL breaks if a parent project also uses it,
since gradle won't let you build the
We have to switch to using the old buildscript plugin system,
so that it can be used both as a subproject and as an independent
project.

Move 'rawComputeVersionSignature' into it's own file,
so that parent projects can use it too.